requirements.txt
pandas
openpyxl
requests
python-dotenv
beautifulsoup4
lxml
openai
httpx
pyarrow